+++ packages/config/tailwind-config/content.ts
@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/no-unsafe-return */
-
 import type { Config } from 'tailwindcss';
 
 import path from 'node:path';
@@ -19,9 +17,12 @@ const HEROUI_THEME_USED_COMPONENT_PATHS = [
   'badge',
   'breadcrumbs',
   'button',
+  'calendar',
   'card',
   'checkbox',
   'chip',
+  'date-input',
+  'date-picker',
   'divider',
   'drawer',
   'dropdown',
@@ -41,7 +42,6 @@ const HEROUI_THEME_USED_COMPONENT_PATHS = [
   'tooltip',
   // theme is not required
 ].map((c) => {
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-return
   return path.resolve(
     __dirname,
     `./node_modules/@heroui/theme/dist/components/${c}.js`,