+++ packages/uniform/src/Time/index.ts
@@ -1,0 +1,2 @@
+export { default as Time } from './Time';
+export type { TimeProps } from './Time';